Cleaned the code up. Added some more functionality
TODO:
Need to Work on the Insertion of a student Page, the function is there, just
don't as of yet have it implemented in it's own page.
Comment the crap out of the code. This is very hard to follow code at a glance,
Try and clean up the points that need to be
cleaned up,
Remove dummy variables,
Fix up the Queries where they can be.
Remove the Lol tag.
Don't think I will do the full graduateable students list.